[Refactor] SSD_Shell.cpp/.h 분리 1. 변경점 - Header/cpp 파일 분리 - 일부 TC 파일 분리 (from  main.cpp)
2. 검증
[X] Build
[X] Regression Test Pass
[ ] 해당 기능 TC
