--- TeamBest_SSD_Shell/ISSD_Shell.h
@@ -1,8 +1,0 @@
-﻿#pragma once
-
-#include <iostream>
-
-class IShell {
-public:
-	virtual void Run() = 0;
-};
