+++ TeamBest_SSD_Shell/TeamBest_SSD_Shell.vcxproj.filters
@@ -27,17 +27,34 @@
     <ClCompile Include="test_ssd.cpp">
       <Filter>테스트 파일</Filter>
     </ClCompile>
-    <ClCompile Include="ISSD_Shell.h">
-      <Filter>헤더 파일</Filter>
-    </ClCompile>
     <ClCompile Include="SSD_Shell.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="test_ssd_shell.cpp">
       <Filter>테스트 파일</Filter>
     </ClCompile>
+    <ClCompile Include="tesh_shellProgram.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gtest\gtest-all.cc">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gmock\gmock-all.cc">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_ssd_shell1.cpp">
+      <Filter>테스트 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_ssd_shell2.cpp">
+      <Filter>테스트 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="SSD_Shell.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
