+++ TeamBest_SSD_Shell/test_ssd_shell1.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include "gmock/gmock.h"
-#include "SSD_Shell.cpp"
+#include "SSD_Shell.h"
 
 using namespace testing;
